[REF] event, website_event: extracte event_question from website_event
In order to use custom questions on events and event registrations
without having to install the website_* module suite, this commit
extracts this functionality from the website_event module to event.
