+++ addons/event/__manifest__.py
@@ -37,7 +37,9 @@ Key Features
         'views/res_config_settings_views.xml',
         'views/event_templates.xml',
         'views/res_partner_views.xml',
-        'views/event_tag_views.xml'
+        'views/event_tag_views.xml',
+        'views/event_question_views.xml',
+        'views/event_registration_answer_views.xml',
     ],
     'demo': [
         'data/res_users_demo.xml',
