+++ addons/event/models/__init__.py
@@ -10,3 +10,6 @@ from . import event_ticket
 from . import mail_template
 from . import res_config_settings
 from . import res_partner
+from . import event_question_answer
+from . import event_registration_answer
+from . import event_question
