+++ addons/event/models/event_event.py
@@ -52,6 +52,13 @@ class EventType(models.Model):
                   'template_ref': 'mail.template, %i' % self.env.ref('event.event_reminder').id,
                  })]
 
+    def _default_question_ids(self):
+        return [
+            (0, 0, {'title': _('Name'), 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': _('Email'), 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': _('Phone'), 'question_type': 'phone'}),
+        ]
+
     name = fields.Char('Event Template', required=True, translate=True)
     note = fields.Html(string='Note')
     sequence = fields.Integer(default=10)
@@ -73,6 +80,9 @@ class EventType(models.Model):
     # ticket reports
     ticket_instructions = fields.Html('Ticket Instructions', translate=True,
         help="This information will be printed on your tickets.")
+    question_ids = fields.One2many(
+        'event.question', 'event_type_id', default=_default_question_ids,
+        string='Questions', copy=True)
 
     @api.depends('has_seats_limitation')
     def _compute_seats_max(self):
@@ -113,6 +123,9 @@ class EventEvent(models.Model):
     @api.model
     def _lang_get(self):
         return self.env['res.lang'].get_installed()
+
+    def _default_question_ids(self):
+        return self.env['event.type']._default_question_ids()
 
     name = fields.Char(string='Event', translate=True, required=True)
     note = fields.Html(string='Note', store=True, compute="_compute_note", readonly=False)
@@ -230,11 +243,55 @@ class EventEvent(models.Model):
     ticket_instructions = fields.Html('Ticket Instructions', translate=True,
         compute='_compute_ticket_instructions', store=True, readonly=False,
         help="This information will be printed on your tickets.")
+    # questions
+    question_ids = fields.One2many(
+        'event.question', 'event_id', 'Questions', copy=True,
+        compute='_compute_question_ids', readonly=False, store=True)
+    general_question_ids = fields.One2many('event.question', 'event_id', 'General Questions',
+                                           domain=[('once_per_order', '=', True)])
+    specific_question_ids = fields.One2many('event.question', 'event_id', 'Specific Questions',
+                                            domain=[('once_per_order', '=', False)])
 
     def _compute_use_barcode(self):
         use_barcode = self.env['ir.config_parameter'].sudo().get_param('event.use_event_barcode') == 'True'
         for record in self:
             record.use_barcode = use_barcode
+
+    @api.depends('event_type_id')
+    def _compute_question_ids(self):
+        """ Update event questions from its event type. Depends are set only on
+        event_type_id itself to emulate an onchange. Changing event type content
+        itself should not trigger this method.
+
+        When synchronizing questions:
+
+          * lines with no registered answers are removed;
+          * type lines are added;
+        """
+        if self._origin.question_ids:
+            # lines to keep: those with already given answers
+            questions_tokeep_ids = self.env['event.registration.answer'].search(
+                [('question_id', 'in', self._origin.question_ids.ids)]
+            ).question_id.ids
+        else:
+            questions_tokeep_ids = []
+        for event in self:
+            if not event.event_type_id and not event.question_ids:
+                event.question_ids = self._default_question_ids()
+                continue
+
+            if questions_tokeep_ids:
+                questions_toremove = event._origin.question_ids.filtered(
+                    lambda question: question.id not in questions_tokeep_ids)
+                command = [(3, question.id) for question in questions_toremove]
+            else:
+                command = [(5, 0)]
+            event.question_ids = command
+
+            # copy questions so changes in the event don't affect the event type
+            event.question_ids += event.event_type_id.question_ids.copy({
+                'event_type_id': False,
+            })
 
     @api.depends('stage_id', 'kanban_state')
     def _compute_kanban_state_label(self):
