+++ addons/website_event/models/__init__.py
@@ -2,10 +2,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import event_event
-from . import event_question
-from . import event_question_answer
 from . import event_registration
-from . import event_registration_answer
 from . import event_tag_category
 from . import event_tag
 from . import event_type
