+++ addons/website_event/security/ir.model.access.csv
@@ -19,5 +19,6 @@ access_website_event_menu_user,website.event.menu.user,model_website_event_menu,
 access_website_visitor_user,website.visitor.user,model_website_visitor,event.group_event_user,1,1,0,0
 access_event_question_public,event.question,event.model_event_question,base.group_public,1,0,0,0
 access_event_question_portal,event.question,event.model_event_question,base.group_portal,1,0,0,0
+access_event_question_user,event.question,event.model_event_question,base.group_user,1,0,0,0
 access_event_question_answer_public,event.question.answer,event.model_event_question_answer,base.group_public,1,0,0,0
 access_event_question_answer_portal,event.question.answer,event.model_event_question_answer,base.group_portal,1,0,0,0
